Update main.md
Updated sentence to read better. Replaced commas with full stops while there.

I re-read the existing sentence 5 time and just could not get what it was trying to tell me because I did not know what a "bypass" was. In my experience not a common way to describe that x = y scenario.